+++ app/src/main/java/com/example/mediaapp/SearchViewModelProviderFactory.kt
@@ -7,18 +7,18 @@ import androidx.lifecycle.ViewModel
 import androidx.savedstate.SavedStateRegistryOwner
 import java.lang.IllegalArgumentException
 
-class SearchViewModelProviderFactory (
+class SearchViewModelProviderFactory(
     private val searchRepository: VideoRepository,
     owner: SavedStateRegistryOwner,
-    defaultArgs : Bundle? = null
-): AbstractSavedStateViewModelFactory(owner,defaultArgs) {
+    defaultArgs: Bundle? = null
+) : AbstractSavedStateViewModelFactory(owner, defaultArgs) {
     override fun <T : ViewModel> create(
         key: String,
         modelClass: Class<T>,
         handle: SavedStateHandle
     ): T {
-        if (modelClass.isAssignableFrom(SearchViewModel::class.java)){
-            return SearchViewModel(searchRepository,handle) as T
+        if (modelClass.isAssignableFrom(SearchViewModel::class.java)) {
+            return SearchViewModel(searchRepository, handle) as T
         }
         throw IllegalArgumentException("View Model class not found")
     }
